requirements.txt
commonroad_vehicle_models>=2.0.0
matplotlib>=2.2.2
networkx>=2.2
numpy>=1.13
methodtools
omegaconf>=2.1.1
pytest>=6.2.5
scipy>=1.5.2
sphinx>=2.2.1
git+https://gitlab.lrz.de/av2.0/commonroad/commonroad-route-planner.git@develop
commonroad-io>=2021.4
commonroad-drivability-checker>=2021.4
numba
commonroad-helper-functions
wale-net
ruamel.yaml















